rename detail_view.ts -> jsonViewer.ts
